Add Euribor functionality and help screen to TUI application
- Implemented Euribor fetching and management in the TUI app.
- Added fields for Euribor date, rate, and buttons for fetching and adding Euribor points.
- Enhanced the calculation logic to incorporate fetched Euribor rates.
- Introduced a help screen with keyboard shortcuts and usage instructions.
- Updated UI to display new Euribor-related fields and actions.
- Added tests for Euribor functionality in the mortgage GUI.
