+++ mortgage_gui/src/chart.rs
@@ -1,13 +1,23 @@
 use mortgage_core::models::LoanResult;
 use plotters::prelude::*;
+use plotters::style::{FontDesc, FontFamily, FontStyle};
 
-pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> String {
+fn default_font(size: f64) -> FontDesc<'static> {
+    FontDesc::new(FontFamily::SansSerif, size, FontStyle::Normal)
+}
+
+pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> Result<String, String> {
     let mut svg_data = String::new();
     {
         let root = SVGBackend::with_string(&mut svg_data, (900, 500)).into_drawing_area();
-        root.fill(&WHITE).unwrap();
+        root.fill(&WHITE)
+            .map_err(|e| format!("Fill error: {}", e))?;
 
         let n = result.payments.len();
+        if n == 0 {
+            return Err("No payments to chart".to_string());
+        }
+
         let max_y = result
             .payments
             .iter()
@@ -15,15 +25,22 @@ pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> String {
             .fold(0.0, f64::max)
             * 1.1;
 
+        if max_y <= 0.0 {
+            return Err("Invalid chart data: max_y is zero".to_string());
+        }
+
         let mut chart = ChartBuilder::on(&root)
-            .caption("Principal vs Interest (Stacked)", ("sans-serif", 28))
+            .caption("Principal vs Interest (Stacked)", default_font(28.0))
             .margin(10)
             .x_label_area_size(30)
             .y_label_area_size(60)
             .build_cartesian_2d(0.0..n as f64, 0.0..max_y)
-            .unwrap();
+            .map_err(|e| format!("Chart build error: {}", e))?;
 
-        chart.configure_mesh().draw().unwrap();
+        chart
+            .configure_mesh()
+            .draw()
+            .map_err(|e| format!("Mesh draw error: {}", e))?;
 
         let bar_width = 0.8;
         for (idx, p) in result.payments.iter().enumerate() {
@@ -36,7 +53,7 @@ pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> String {
                     ],
                     GREEN.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw principal error: {}", e))?;
             chart
                 .draw_series(std::iter::once(Rectangle::new(
                     [
@@ -45,7 +62,7 @@ pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> String {
                     ],
                     RED.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw interest error: {}", e))?;
         }
 
         if let Some(cross_idx) = result.principal_exceeds_interest_at {
@@ -59,34 +76,39 @@ pub fn generate_stacked_bar_chart_svg(result: &LoanResult) -> String {
                     6,
                     BLUE.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw circle error: {}", e))?;
             chart
                 .draw_series(std::iter::once(Text::new(
                     format!("Cross #{} ({})", cross_idx + 1, cross_payment.date),
                     (cross_idx as f64, max_y * 0.92),
-                    ("sans-serif", 12).into_font().color(&BLUE),
+                    default_font(12.0).color(&BLUE),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw text error: {}", e))?;
         }
 
         chart
             .configure_series_labels()
             .border_style(BLACK)
             .draw()
-            .unwrap();
+            .map_err(|e| format!("Labels draw error: {}", e))?;
     }
-    svg_data
+    Ok(svg_data)
 }
 
-pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
+pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Result<Vec<u8>, String> {
     use image::ImageEncoder;
     let (width, height) = (900u32, 500u32);
     let mut raw = vec![0u8; (width * height * 3) as usize];
     {
         let root = BitMapBackend::with_buffer(&mut raw, (width, height)).into_drawing_area();
-        root.fill(&WHITE).unwrap();
+        root.fill(&WHITE)
+            .map_err(|e| format!("Fill error: {}", e))?;
 
         let n = result.payments.len();
+        if n == 0 {
+            return Err("No payments to chart".to_string());
+        }
+
         let max_y = result
             .payments
             .iter()
@@ -94,15 +116,22 @@ pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
             .fold(0.0, f64::max)
             * 1.1;
 
+        if max_y <= 0.0 {
+            return Err("Invalid chart data: max_y is zero".to_string());
+        }
+
         let mut chart = ChartBuilder::on(&root)
-            .caption("Principal vs Interest (Stacked)", ("sans-serif", 28))
+            .caption("Principal vs Interest (Stacked)", default_font(28.0))
             .margin(10)
             .x_label_area_size(30)
             .y_label_area_size(60)
             .build_cartesian_2d(0.0..n as f64, 0.0..max_y)
-            .unwrap();
+            .map_err(|e| format!("Chart build error: {}", e))?;
 
-        chart.configure_mesh().draw().unwrap();
+        chart
+            .configure_mesh()
+            .draw()
+            .map_err(|e| format!("Mesh draw error: {}", e))?;
 
         let bar_width = 0.8;
         for (idx, p) in result.payments.iter().enumerate() {
@@ -115,7 +144,7 @@ pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
                     ],
                     GREEN.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw principal error: {}", e))?;
             chart
                 .draw_series(std::iter::once(Rectangle::new(
                     [
@@ -124,7 +153,7 @@ pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
                     ],
                     RED.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw interest error: {}", e))?;
         }
 
         if let Some(cross_idx) = result.principal_exceeds_interest_at {
@@ -138,21 +167,21 @@ pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
                     6,
                     BLUE.filled(),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw circle error: {}", e))?;
             chart
                 .draw_series(std::iter::once(Text::new(
                     format!("Cross #{} ({})", cross_idx + 1, cross_payment.date),
                     (cross_idx as f64, max_y * 0.92),
-                    ("sans-serif", 12).into_font().color(&BLUE),
+                    default_font(12.0).color(&BLUE),
                 )))
-                .unwrap();
+                .map_err(|e| format!("Draw text error: {}", e))?;
         }
 
         chart
             .configure_series_labels()
             .border_style(BLACK)
             .draw()
-            .unwrap();
+            .map_err(|e| format!("Labels draw error: {}", e))?;
     }
 
     let mut png_buf = Vec::new();
@@ -160,18 +189,23 @@ pub fn generate_stacked_bar_chart_png(result: &LoanResult) -> Vec<u8> {
         let mut cursor = std::io::Cursor::new(&mut png_buf);
         image::codecs::png::PngEncoder::new(&mut cursor)
             .write_image(&raw, width, height, image::ColorType::Rgb8)
-            .unwrap();
+            .map_err(|e| format!("PNG encode error: {}", e))?;
     }
-    png_buf
+    Ok(png_buf)
 }
 
-pub fn generate_balance_line_chart_svg(result: &LoanResult) -> String {
+pub fn generate_balance_line_chart_svg(result: &LoanResult) -> Result<String, String> {
     let mut svg_data = String::new();
     {
         let root = SVGBackend::with_string(&mut svg_data, (900, 500)).into_drawing_area();
-        root.fill(&WHITE).unwrap();
+        root.fill(&WHITE)
+            .map_err(|e| format!("Fill error: {}", e))?;
 
         let n = result.payments.len();
+        if n == 0 {
+            return Err("No payments to chart".to_string());
+        }
+
         let max_y = result
             .payments
             .iter()
@@ -179,15 +213,22 @@ pub fn generate_balance_line_chart_svg(result: &LoanResult) -> String {
             .fold(0.0, f64::max)
             * 1.1;
 
+        if max_y <= 0.0 {
+            return Err("Invalid chart data: max_y is zero".to_string());
+        }
+
         let mut chart = ChartBuilder::on(&root)
-            .caption("Remaining Balance Over Time", ("sans-serif", 28))
+            .caption("Principal vs Interest (Stacked)", default_font(28.0))
             .margin(10)
             .x_label_area_size(30)
             .y_label_area_size(60)
             .build_cartesian_2d(0.0..n as f64, 0.0..max_y)
-            .unwrap();
+            .map_err(|e| format!("Chart build error: {}", e))?;
 
-        chart.configure_mesh().draw().unwrap();
+        chart
+            .configure_mesh()
+            .draw()
+            .map_err(|e| format!("Mesh draw error: {}", e))?;
 
         let points: Vec<(f64, f64)> = result
             .payments
@@ -196,24 +237,31 @@ pub fn generate_balance_line_chart_svg(result: &LoanResult) -> String {
             .map(|(i, p)| (i as f64, p.remaining_balance))
             .collect();
 
-        chart.draw_series(LineSeries::new(points, &BLUE)).unwrap();
+        chart
+            .draw_series(LineSeries::new(points, &BLUE))
+            .map_err(|e| format!("Draw line error: {}", e))?;
 
         chart
             .configure_series_labels()
             .border_style(BLACK)
             .draw()
-            .unwrap();
+            .map_err(|e| format!("Labels draw error: {}", e))?;
     }
-    svg_data
+    Ok(svg_data)
 }
 
-pub fn generate_overlay_chart_svg(result: &LoanResult) -> String {
+pub fn generate_overlay_chart_svg(result: &LoanResult) -> Result<String, String> {
     let mut svg_data = String::new();
     {
         let root = SVGBackend::with_string(&mut svg_data, (900, 500)).into_drawing_area();
-        root.fill(&WHITE).unwrap();
+        root.fill(&WHITE)
+            .map_err(|e| format!("Fill error: {}", e))?;
 
         let n = result.payments.len();
+        if n == 0 {
+            return Err("No payments to chart".to_string());
+        }
+
         let max_payment = result
             .payments
             .iter()
@@ -228,15 +276,22 @@ pub fn generate_overlay_chart_svg(result: &LoanResult) -> String {
             * 1.1;
         let max_y = max_payment.max(max_balance);
 
+        if max_y <= 0.0 {
+            return Err("Invalid chart data: max_y is zero".to_string());
+        }
+
         let mut chart = ChartBuilder::on(&root)
-            .caption("Amortization Overview", ("sans-serif", 28))
+            .caption("Amortization Overview", default_font(28.0))
             .margin(10)
             .x_label_area_size(30)
             .y_label_area_size(60)
             .build_cartesian_2d(0.0..n as f64, 0.0..max_y)
-            .unwrap();
+            .map_err(|e| format!("Chart build error: {}", e))?;
 
-        chart.configure_mesh().draw().unwrap();
+        chart
+            .configure_mesh()
+            .draw()
+            .map_err(|e| format!("Mesh draw error: {}", e))?;
 
         let principal_pts: Vec<(f64, f64)> = result
             .payments
@@ -259,19 +314,19 @@ pub fn generate_overlay_chart_svg(result: &LoanResult) -> String {
 
         chart
             .draw_series(LineSeries::new(principal_pts, &GREEN))
-            .unwrap();
+            .map_err(|e| format!("Draw principal line error: {}", e))?;
         chart
             .draw_series(LineSeries::new(interest_pts, &RED))
-            .unwrap();
+            .map_err(|e| format!("Draw interest line error: {}", e))?;
         chart
             .draw_series(LineSeries::new(balance_pts, &BLUE))
-            .unwrap();
+            .map_err(|e| format!("Draw balance line error: {}", e))?;
 
         chart
             .configure_series_labels()
             .border_style(BLACK)
             .draw()
-            .unwrap();
+            .map_err(|e| format!("Labels draw error: {}", e))?;
     }
-    svg_data
+    Ok(svg_data)
 }
